// src/public/gamesCollection.d.ts
export const GamesCollection: {
    id: number;
    title_ru: string;
    title_en: string;
    subtittle_ru: string;
    subtittle_en: string;
    description_ru: string;
    description_en: string;
    price: number;
    discountPercentage: number;
    rating: number;
    stock: number;
    GameTime: number[];
    age: number[];
    gamers: number[];
    brand: string;
    category_ru: string[];
    category_en: string[];
    thumbnail: any;
    images: any[];
}[];
//# sourceMappingURL=gamesCollection.d.ts.map